Fjern unødvendig funksjonalitet fra temabygger
