+++ portal/src/app/(frontend)/temabygger/layout.tsx
@@ -1,6 +1,4 @@
 import "./global.scss";
-import { ThemeBuilderProvider } from "@/components/theme-builder/ThemeBuilderProvider";
-import { tokensFromSchema } from "@/components/theme-builder/tokens";
 
 type ThemeBuilderLayoutProps = {
     children: React.ReactNode;
@@ -9,9 +7,5 @@ type ThemeBuilderLayoutProps = {
 export default function ThemeBuilderLayout({
     children,
 }: ThemeBuilderLayoutProps) {
-    return (
-        <ThemeBuilderProvider baseTokens={tokensFromSchema()}>
-            {children}
-        </ThemeBuilderProvider>
-    );
+    return children;
 }
